requirements.txt
Django
django-heroku
dj-database-url
eventlet
gunicorn
python-socketio
psycopg2-binary
Celery
redis